pyproject.toml
[project]
name = "suggestions-version-4"
version = "4.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiobotocore>=2.22.0",
    "arrow>=1.4.0",
    "fastnanoid>=0.4.1",
    "freezegun>=1.5.2",
    "function-cooldowns>=2.2.0",
    "hikari-lightbulb>=3.0.0a25",
    "hikari[speedups]>=2.3.3",
    "httpx-oauth>=0.16.1",
    "httpx-retries>=0.4.5",
    "infisicalsdk>=1.0.8",
    "litestar[cryptography,opentelemetry]>=2.19.0",
    "logoo>=1.6.0",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-exporter-otlp-proto-http>=1.39.1",
    "opentelemetry-sdk>=1.39.1",
    "orjson>=3.10.18",
    "piccolo-admin>=1.12.0",
    "piccolo-api[authenticator,pynacl]>=1.7.0",
    "piccolo[all,postgres]>=1.26.0",
    "python-dotenv>=1.1.0",
    "redis[hiredis]>=6.2.0",
    "saq[hiredis,web]>=0.26.0",
    "skelmis-commons>=1.6.1",
    "tenacity>=9.1.2",
    "uvloop>=0.22.1",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "fakeredis>=2.27.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
]
